check balanced string

+++ Arrays & Strings/Leetcode_3340.cpp
@@ -3,7 +3,7 @@ public:
     bool isBalanced(string num) {
         int oddSum = 0;
         int evenSum = 0;
-        for(int i=0; i<num.length(); ++i){
+        for(size_t i=0; i<num.length(); ++i){
             int digit = num[i] - '0';
             if(i & 1)
                 oddSum += digit;
